# api/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API to manage Event.",
        "title": "middleware/project",
        "contact": {
            "name": "Baye serigne Seck.",
            "email": "Baye_Serigne.SECK@etu.uca.fr"
        },
        "version": "1.0.0"
    },
    "basePath": "/",
    "paths": {
        "/events": {
            "get": {
                "description": "Récupère la liste complète des événements disponibles",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Events"
                ],
                "summary": "Récupère tous les événements",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Event"
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur lors de la récupération des événements",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/events/resource/{resource_id}": {
            "get": {
                "description": "Récupère les événements associés à une ressource spécifique",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Events"
                ],
                "summary": "Récupère les événements d'une ressource",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID de la ressource",
                        "name": "resource_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Event"
                            }
                        }
                    },
                    "400": {
                        "description": "ID invalide",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Erreur lors de la récupération des événements",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Event": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "dtend": {
                    "type": "string"
                },
                "dtstart": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "resource_id": {
                    "type": "string"
                },
                "summary": {
                    "type": "string"
                },
                "uid": {
                    "type": "string"
                }
            }
        }
    }
}